% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adalasso.R
\name{adalasso}
\alias{adalasso}
\title{Implement Adaptive Lasso via glmnet}
\usage{
adalasso(
  x,
  y,
  lambda,
  lambda_init = NULL,
  gamma = 1,
  intercept = TRUE,
  scale_x = FALSE,
  model = NULL
)
}
\arguments{
\item{x}{Predictor matrix (n-by-p matrix)}

\item{y}{Response variable}

\item{lambda}{Shrinkage tuning parameter for the adaptive step}

\item{lambda_init}{Shrinkage tuning parameter for the initial step Lasso estimation (If applicable)}

\item{gamma}{Parameter controlling the inverse of first step estimate}

\item{intercept}{A boolean: include an intercept term or not}

\item{scale_x}{A boolean: standardize the design matrix or not}

\item{model}{A string: model in the initial estimation, "lasso", "ridge", or "ols". Default is "lasso" for high-dimensional case and "ols" for low-dimensional case.}
}
\value{
A list contains estimated intercept and slope
\item{ahat}{Estimated intercept}
\item{bhat}{Estimated slope}
}
\description{
Estimation function. Tuning parameter inputs needed.\cr
Incorporates both high-dim(Lasso as initial estimator) and low-dim (OLS as initial estimator).
}
\examples{
\dontrun{
x <- matrix(rnorm(50 * 5), 50, 5)
y <- rnorm(50)
adalasso(x, y, lambda = 0.1)
}
}
\keyword{internal}
