% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_functions.R
\name{AR.URtest}
\alias{AR.URtest}
\title{Estimation of AR coefficients and do unit root tests}
\usage{
AR.URtest(Data, Roll.Window)
}
\description{
Estimation of AR coefficients and do unit root tests
}
