% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{autoplot.lasforecast_roll}
\alias{autoplot.lasforecast_roll}
\title{Autoplot method for lasforecast_roll objects}
\usage{
\method{autoplot}{lasforecast_roll}(object, methods_to_plot = NULL, date_range = NULL, ...)
}
\arguments{
\item{object}{An object of class "lasforecast_roll" generated by \code{\link{roll_predict}}.}

\item{methods_to_plot}{Character vector of methods to plot. Default uses all.}

\item{date_range}{Length-2 vector of dates (or strings coercible to Date)
to restrict the plotted time span. Ignored when dates are unavailable.}

\item{...}{Additional arguments.}
}
\description{
Autoplot method for lasforecast_roll objects
}
