% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss.R
\name{bss.bic}
\alias{bss.bic}
\title{Implement best subset selection via Mixed Integer Optimization (MIO)
Select k by BIC}
\usage{
bss.bic(
  y,
  X,
  intercept = TRUE,
  b0 = NULL,
  tau = 2,
  tol = 1e-04,
  MaxIter = 10000,
  polish = TRUE,
  time.limit = 1200,
  RW = TRUE
)
}
\arguments{
\item{y}{forecast target}

\item{X}{predictors}

\item{b0}{initial estimator}

\item{tau}{parameter to obtain bounds}

\item{tol}{precision tolerence}

\item{polish}{whether post-selection polish is conducted}

\item{time.limit}{in seconds}

\item{RW}{consider k equal to 0 or not}
}
\value{
A list contains estimated k and corresponding coefficient
\item{k}{estimated k}
\item{coef}{estimated coefficient}
}
\description{
Make use of Gurobi solver
}
