% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{coef.lasforecast_model}
\alias{coef.lasforecast_model}
\title{Extract coefficients from a lasforecast_model}
\usage{
\method{coef}{lasforecast_model}(object, ...)
}
\arguments{
\item{object}{A \code{lasforecast_model} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Numeric vector of coefficients.
}
\description{
Extract coefficients from a lasforecast_model
}
