% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csr.R
\name{csr.bic}
\alias{csr.bic}
\title{Elliott, Gargano and Timmermann (2013) COMPLETE SUBSET REGRESSIONS with BIC}
\usage{
csr.bic(y, X, C.upper = 5000, intercept = FALSE, RW = TRUE)
}
\arguments{
\item{y}{response variable}

\item{X}{Predictor matrix}

\item{C.upper}{maximum number of subsets to be combined}

\item{intercept}{A boolean: include an intercept term or not}
}
\value{
A List contained the estimated coefficients and forecasts
\item{k.hat}{k chosen by BIC}
\item{coef}{Averaged coefficients corresonding to the chosen k}
}
\description{
Choose tuning parameter k by Bayesian Information Criterion (BIC)
}
