% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csr.R
\name{csr}
\alias{csr}
\title{Elliott, Gargano and Timmermann (2013) COMPLETE SUBSET REGRESSIONS}
\usage{
csr(y, X, k, C.upper = 5000, intercept = FALSE)
}
\arguments{
\item{y}{response variable}

\item{X}{Predictor matrix}

\item{k}{subset size}

\item{C.upper}{maximum number of subsets to be combined}

\item{intercept}{A boolean: include an intercept term or not}
}
\value{

}
\description{
Incorporates techniques in section 3.4 This step aims to reduce the
computation burden when K is large, but it is not feasible when K is too
large using standard R functions For example, when K = 50, k = 25, the vector
1:choose(50,25) consumes 941832.4 Gb memory, which is an astronomical number
Future worke: see Boot and Nibbering (2019) Random subspace method.
}
