% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2relax.R
\name{est_sigma_mat}
\alias{est_sigma_mat}
\title{Estimate the sample covariance matrix}
\usage{
est_sigma_mat(y, x)
}
\arguments{
\item{y}{forecast target}

\item{x}{forecasts to be combined}
}
\value{
sigma_mat: The estimated sample covariance matrix
}
\description{
Consider different methods later on.
}
