% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{find_tau_max}
\alias{find_tau_max}
\title{Find the largest tau for L2-Relaxation}
\usage{
find_tau_max(sigma_mat)
}
\arguments{
\item{sigma_mat}{Sample covariance matrix}
}
\value{
smallest tau corresponds to equal-weight
}
\description{
This function finds the smallest tau for L2-Relaxation such that
equal-weight solves the forecast combination optimization.
}
