% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{find_tau_max_reg}
\alias{find_tau_max_reg}
\title{Find the minimum tau such that equal weight solve the l_2 relaxation problem}
\usage{
find_tau_max_reg(y, X, solver = "CVXR", intercept = TRUE, tol = 1e-06)
}
\arguments{
\item{solver}{The solver to use; "Rmosek" or "CVXR"}

\item{tol}{Tolerance for the solver}
}
\description{
Find the minimum tau such that equal weight solve the l_2 relaxation problem
}
