% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiased_ivx_est.R
\name{generate_iv}
\alias{generate_iv}
\title{Self-generated IVs}
\usage{
generate_iv(d, n, a, c_z = 5)
}
\arguments{
\item{d}{Vector of regressor values}

\item{n}{Sample size}

\item{a}{Parameter in constructing IV}

\item{c_z}{Parameter in constructing IV}
}
\value{
Instrumental variable vector of length n-1
}
\description{
Self-generated IVs
}
\keyword{internal}
