% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lasforecast.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Generate ggplot2 default color palette}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of colors.}
}
\value{
Character vector of hex color codes.
}
\description{
Replicates the default ggplot2 color palette (equally spaced hues).
}
\keyword{internal}
