% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2relax.R
\name{l2_relax}
\alias{l2_relax}
\title{l2-relaxation forecast combination
A wrapper function with parameter tuning and estimation.}
\usage{
l2_relax(y, x, tau, solver = "CVXR", tol = 1e-08)
}
\arguments{
\item{y}{foreccasting target}

\item{x}{forecasts to be combined}

\item{tau}{The regularization parameter}

\item{solver}{The solver to use; "Rmosek" or "CVXR"}

\item{tol}{Tolerance for the solver}
}
\value{
A list
}
\description{
Details refer to Shi, Su and Xie (2022)
  "l_2-relaxation:
     With applications to forecast combination and portfolio analysis"
}
