% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2relax.R
\name{l2_relax_reg_opt}
\alias{l2_relax_reg_opt}
\title{Solve L2-relaxation for a regression problem}
\usage{
l2_relax_reg_opt(y, X, tau, intercept = TRUE, solver = "CVXR", tol = 1e-06)
}
\arguments{
\item{tau}{The regularization parameter}

\item{solver}{The solver to use; "Rmosek" or "CVXR"}

\item{tol}{Tolerance for the solver}
}
\description{
Solve L2-relaxation for a regression problem
}
