% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_weight.R
\name{lasso_weight_opt}
\alias{lasso_weight_opt}
\title{Implement weighted Lasso estimation (second step of adaptive lasso) via optimization solver}
\usage{
lasso_weight_opt(x, y, lambda, w = NULL, intercept = TRUE,
  scalex = FALSE, solver = "CVXR", rtol = 1e-08, verb = 0)
}
\arguments{
\item{x}{Predictor matrix (n-by-p matrix)}

\item{y}{Response variable}

\item{lambda}{Shrinkage tuning parameter}

\item{w}{weights}

\item{intercept}{A boolean: include an intercept term or not}

\item{scalex}{A boolean: standardize the design matrix or not}

\item{solver}{indicate solver in use, c("CVXR", "MOSEK")}
}
\value{
A list contains estimated intercept and slope
\item{ahat}{Estimated intercept}
\item{bhat}{Estimated slope}
}
\description{
Estimation function. Tuning parameter inputs needed.\cr
}
\examples{
lasso_weight_opt(x,y)
}
