% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrcov.R
\name{lrcov_est}
\alias{lrcov_est}
\title{Long-run covariance estimator}
\usage{
lrcov_est(x, y = NULL, type = 0)
}
\arguments{
\item{x}{A numeric matrix or vector}

\item{y}{A numeric matrix or vector (optional)}

\item{type}{Integer 0 or 1. 0 for long-run covariance, 1 for one-sided.}
}
\description{
Use bartlett kernel to estimate long-run covariance where the bandwidth
determined Andrews (1991, Eq. (6.2) and (6.4)).
Part of the function adapted from the "bwAndrews" function in
"sandwich" package
}
\keyword{internal}
