% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{ols}
\alias{ols}
\title{Estimate OLS Model}
\usage{
ols(x, y = NULL, intercept = TRUE, date_col = NULL, ...)
}
\arguments{
\item{x}{Predictor data. Can be a matrix, data.frame, tibble, or tsibble.}

\item{y}{Response vector, or column name string when \code{x} is a data frame.}

\item{intercept}{Logical}

\item{date_col}{Optional date column name (for data.frame/tibble input).}

\item{...}{Additional arguments (ignored).}
}
\value{
A list with class "lasforecast_model".
}
\description{
Fits an OLS model.
}
