% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_coef}
\alias{plot_coef}
\title{Coefficient plot}
\usage{
plot_coef(
  coef_est,
  dates,
  pt_num = 4,
  col_vec = NULL,
  line_size = rep(0.75, length(coef_est)),
  alpha_size = rep(0.75, length(coef_est)),
  xlab = NULL,
  ylab = NULL,
  num_col = 1
)
}
\arguments{
\item{coef_est}{estimated slope by m different methods, list of length m with each element n-by-p matrix\cr
names of list well defined}

\item{dates}{Date vector of class "Date".}

\item{pt_num}{Number date points (pt_num + 1) wanted on the x-axis}

\item{col_vec}{A vector indicates colors of different trends.}

\item{line_size}{line size for each trend}

\item{alpha_size}{degree of appearance}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{num_col}{number of columns in facet wrap}
}
\value{
A ggplot output
}
\description{
Coefficient plot
}
