% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postlasso.R
\name{post_lasso}
\alias{post_lasso}
\title{post-selection estimation}
\usage{
post_lasso(x, y, coef_est, intercept = TRUE, scale_x = FALSE)
}
\arguments{
\item{x}{Predictor matrix (n-by-p matrix)}

\item{y}{Response variable}

\item{coef_est}{shrinkage estimation results}

\item{intercept}{A boolean: include an intercept term or not}

\item{scale_x}{A boolean: standardize the design matrix or not}
}
\value{
A list contains estimated intercept and slope
\item{ahat}{Estimated intercept}
\item{bhat}{Estimated slope}
}
\description{
post-selection estimation
}
\examples{
\dontrun{
x <- matrix(rnorm(50 * 5), 50, 5)
y <- rnorm(50)
coef_est <- c(0, 1, 0, 0, -0.5, 0)
post_lasso(x, y, coef_est)
}

}
\keyword{internal}
