% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{predict.lasforecast_model}
\alias{predict.lasforecast_model}
\title{Predict method for LasForecast models}
\usage{
\method{predict}{lasforecast_model}(object, newx, ...)
}
\arguments{
\item{object}{Model object}

\item{newx}{New data matrix}

\item{...}{Additional arguments}
}
\description{
Predict method for LasForecast models
}
