% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.lasforecast_backtest}
\alias{print.lasforecast_backtest}
\title{Print method for lasforecast_backtest objects}
\usage{
\method{print}{lasforecast_backtest}(x, ...)
}
\arguments{
\item{x}{A \code{lasforecast_backtest} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for lasforecast_backtest objects
}
