% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.lasforecast_debias_ivx}
\alias{print.lasforecast_debias_ivx}
\title{Print method for lasforecast_debias_ivx objects}
\usage{
\method{print}{lasforecast_debias_ivx}(x, ...)
}
\arguments{
\item{x}{A \code{lasforecast_debias_ivx} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for lasforecast_debias_ivx objects
}
