% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.lasforecast_model}
\alias{print.lasforecast_model}
\title{Print method for lasforecast_model objects}
\usage{
\method{print}{lasforecast_model}(x, ...)
}
\arguments{
\item{x}{A \code{lasforecast_model} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for lasforecast_model objects
}
