% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.lasforecast_roll}
\alias{print.lasforecast_roll}
\title{Print method for lasforecast_roll objects}
\usage{
\method{print}{lasforecast_roll}(x, ...)
}
\arguments{
\item{x}{A \code{lasforecast_roll} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for lasforecast_roll objects
}
