% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.summary.lasforecast_debias_ivx}
\alias{print.summary.lasforecast_debias_ivx}
\title{Print method for summary.lasforecast_debias_ivx objects}
\usage{
\method{print}{summary.lasforecast_debias_ivx}(x, ...)
}
\arguments{
\item{x}{A \code{summary.lasforecast_debias_ivx} object.}

\item{...}{Additional arguments passed to \code{\link[stats]{printCoefmat}}.}
}
\description{
Print method for summary.lasforecast_debias_ivx objects
}
