% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replasso.R
\name{replasso}
\alias{replasso}
\title{Implement repeated Lasso estimation}
\usage{
replasso(x, y, b.first, lambda, gamma = 1, intercept = TRUE,
  scalex = FALSE, solver = "CVXR")
}
\arguments{
\item{x}{Predictor matrix (n-by-p matrix)}

\item{y}{Response variable}

\item{b.first}{First step adaptive lasso estimates}

\item{lambda}{Shrinkage tuning parameter}

\item{gamma}{Parameter controlling the inverse of first step estimate}

\item{intercept}{A boolean: include an intercept term or not}

\item{solver}{indicate solver in use, c("CVXR", "MOSEK")}
}
\value{
A list contains estimated intercept and slope
\item{ahat}{Estimated intercept}
\item{bhat}{Estimated slope}
}
\description{
Estimation function. Tuning parameter inputs needed. First step adalasso estimate needed.\cr
}
\examples{
replasso(x,y, b.first,lambda)
}
