% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_predict_l2relax.R
\name{roll_predict_l2relax}
\alias{roll_predict_l2relax}
\title{Rolling window forecast}
\usage{
roll_predict_l2relax(
  x,
  y,
  roll_window,
  h = 1,
  k_max = 4,
  m = 5,
  ntau = 100,
  tau_min_ratio = 0.01,
  train_method = "oos",
  solver = "CVXR",
  tol = 1e-07,
  verb = TRUE
)
}
\arguments{
\item{x}{Full sample predictor}

\item{y}{Full sample forecast target}

\item{roll_window}{Length of the rolling window}

\item{h}{forecast horizon}

\item{m}{number of folds}

\item{ntau}{number of tau values}

\item{train_method}{parameter tuning method for L2relax "cv_random", "cv" or "oos"}

\item{solver}{"Rmosek" or "CVXR"}

\item{tol}{tolerance for the solver}

\item{verb}{boolean to control whether print information on screen}

\item{tau.min.ratio}{ratio of the minimum tau in tau.seq over the maximum
(which is the smallest tau such that
equal-weight solves the forecast combination optimization.)}
}
\description{
The function reads data and make forecasts based on linear predictive regression
with diverse methods. It incorporates both short-horizon and long-horizon forecasting.
}
