% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{rw}
\alias{rw}
\title{Estimate Random Walk}
\usage{
rw(x, y = NULL, date_col = NULL, type = "level", ...)
}
\arguments{
\item{x}{Predictor data. Can be a matrix, data.frame, tibble, or tsibble.}

\item{y}{Response vector, or column name string when \code{x} is a data frame.}

\item{date_col}{Optional date column name.}

\item{type}{Type of forecast target y. This is only relevant for RW method.
"diff" for returns/differences and RW predicts zeros; "level" for levels and RW predicts the previous value. Default is "level".}

\item{...}{Additional arguments (ignored).}
}
\description{
Estimate Random Walk
}
