% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{rwwd}
\alias{rwwd}
\title{Estimate RW with Drift}
\usage{
rwwd(x, y = NULL, date_col = NULL, ...)
}
\arguments{
\item{x}{Predictor data. Can be a matrix, data.frame, tibble, or tsibble.}

\item{y}{Response vector, or column name string when \code{x} is a data frame.}

\item{date_col}{Optional date column name.}

\item{...}{Additional arguments (ignored).}
}
\description{
Estimate RW with Drift
}
