% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talasso.R
\name{talasso}
\alias{talasso}
\title{Implement repeated Lasso estimation}
\usage{
talasso(x, y, b_first, lambda, gamma = 1, intercept = TRUE, scale_x = FALSE)
}
\arguments{
\item{x}{Predictor matrix (n-by-p matrix)}

\item{y}{Response variable}

\item{b_first}{First step adaptive lasso estimates}

\item{lambda}{Shrinkage tuning parameter}

\item{gamma}{Parameter controlling the inverse of first step estimate}

\item{intercept}{A boolean: include an intercept term or not}

\item{scale_x}{A boolean: standardize the design matrix or not}
}
\value{
A list contains estimated intercept and slope
\item{ahat}{Estimated intercept}
\item{bhat}{Estimated slope}
}
\description{
Estimation function. Tuning parameter inputs needed. First step adalasso estimate needed.\cr
}
\examples{
\dontrun{
x <- matrix(rnorm(50 * 5), 50, 5)
y <- rnorm(50)
b_first <- rnorm(5)
talasso(x, y, b_first, lambda = 0.1)
}
}
\keyword{internal}
