% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_l2_relax}
\alias{train_l2_relax}
\title{Do parameter tuning for L2-Relaxation}
\usage{
train_l2_relax(
  y,
  x,
  m = 5,
  tau.seq = NULL,
  ntau = 100,
  tau.min.ratio = 0.01,
  train_method = FALSE,
  solver = "Rmosek",
  tol = 1e-05
)
}
\arguments{
\item{y}{foreccasting target}

\item{x}{forecasts to be combined}

\item{m}{number of folds}

\item{tau.seq}{Sequence of tau values}

\item{ntau}{number of tau values}

\item{tau.min.ratio}{ratio of the minimum tau in tau.seq over the maximum 
(which is the smallest tau such that 
equal-weight solves the forecast combination optimization.)}

\item{train_method}{"cv_random", "cv" or "oos"}

\item{solver}{"Rmosek" or "CVXR"}
}
\value{
besttune
}
\description{
Do parameter tuning for L2-Relaxation
}
