% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_replasso}
\alias{train_replasso}
\title{Do parameter tuning for replasso}
\usage{
train_replasso(x, y, b_first, gamma = 1, intercept = TRUE,
  scalex = FALSE, lambda_seq = NULL, nlambda = 100,
  lambda_min_ratio = 1e-04, k = 10, solver = "CVXR")
}
\arguments{
\item{x}{Predictor matrix (n-by-p matrix)}

\item{y}{Response variable}

\item{b_first}{estimated slope from first step alasso}

\item{gamma}{Parameter controlling the inverse of first step estimate. By default = 1.}

\item{intercept}{A boolean: include an intercept term or not}

\item{scalex}{A boolean: standardize the design matrix or not}

\item{lambda_seq}{Candidate sequnece of parameters. If NULL, the function generates the sequnce.}

\item{nlambda}{# of lambdas}

\item{lambda_min_ratio}{# lambda_min_ratio * lambda_max = lambda_min}

\item{k}{k-fold cv if "cv" is chosen}

\item{solver}{"CVXR" or "Rmosek"}

\item{train_method}{"timeslice" or "cv"}
}
\value{
bestTune
}
\description{
Do parameter tuning for replasso
}
\examples{
train_replasso(x,y)
}
