% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_industry_prod}
\alias{US_industry_prod}
\title{Growth Rate of US industry Production Growth.}
\format{
A data frame with 748 rows and 110 variables:
\describe{
  \item{Ind_Growth_Rate}{monthly growth rate of US Industrial Production Index}
  \item{RPI}{real personal income}
  \item{W875RX1}{real personal income ex transfer receipts}
  ...
}
}
\source{
Industrial Production Index \url{https://fred.stlouisfed.org/series/INDPRO}

Remaining Macroeconomic Factors \url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}
}
\usage{
US_industry_prod
}
\description{
A monthly dataset containing the growth rate of US industry production growth rate
and other macroeconomic variables from 1960-01-01 to 2022-04-01
}
\keyword{datasets}
