% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_utils.R
\name{validate_input}
\alias{validate_input}
\title{Validate input dimensions and types}
\usage{
validate_input(x, y)
}
\arguments{
\item{x}{Numeric matrix}

\item{y}{Numeric vector}
}
\description{
Validate input dimensions and types
}
\keyword{internal}
