% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vars_all}
\alias{vars_all}
\title{Predictor Variable Names from FRED-MD}
\format{
A character vector of length 111.
}
\usage{
vars_all
}
\description{
Character vector of 111 predictor variable names from the \code{\link{fredmd}}
dataset. Includes all columns except \code{date} and the target \code{infl_cpi}.
UNRATE is included as a predictor (kept in levels).
}
\keyword{datasets}
