% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adalasso.R
\name{adalasso}
\alias{adalasso}
\title{Implement Adaptive Lasso via glmnet}
\usage{
adalasso(x, y, lambda, lambda.lasso = NULL, gamma = 1,
  intercept = TRUE, scalex = FALSE)
}
\arguments{
\item{x}{Predictor matrix (n-by-p matrix)}

\item{y}{Response variable}

\item{lambda}{Shrinkage tuning parameter for the adaptive step}

\item{gamma}{Parameter controlling the inverse of first step estimate}

\item{intercept}{A boolean: include an intercept term or not}

\item{scalex}{A boolean: standardize the design matrix or not}

\item{lambda_lasso}{Shrinkage tuning parater for the initial step Lasso estimation (If applicable)}
}
\value{
A list contains estimated intercept and slope
\item{ahat}{Estimated intercept}
\item{bhat}{Estimated slope}
}
\description{
Estimation function. Tuning parameter inputs needed.\cr
Incorporates both high-dim(Lasso as initial estimator) and low-dim (OLS as initial estimator).
}
\examples{
adalasso(x,y)
}
