% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiased_ivx_est.R
\name{fit_lasso}
\alias{fit_lasso}
\title{Run Lasso estimation}
\usage{
fit_lasso(
  w,
  y,
  intercept = FALSE,
  standardize = TRUE,
  lambda_choice = NULL,
  lambda_seq = NULL,
  train_method = "timeslice",
  nlambda = 100,
  lambda_min_ratio = 1e-04,
  k = 10,
  initial_window = ceiling(nrow(w) * 0.7),
  horizon = 1,
  fixed_window = TRUE,
  skip = 0
)
}
\arguments{
\item{w}{Matrix of all regressors}

\item{y}{Vector of dependent variable}
}
\value{
coefficients of Lasso regression and residuals
}
\description{
Run Lasso estimation
}
