% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_functions.R
\name{get_lambda_seq}
\alias{get_lambda_seq}
\title{Get the lambda sequence}
\usage{
get_lambda_seq(lambda_max, lambda_min_ratio = 1e-04, nlambda = 100)
}
\arguments{
\item{nlambda}{}
}
\value{
lambda_seq
}
\description{
Get the lambda sequence
}
