% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_functions.R
\name{get_lasso_lambda_max}
\alias{get_lasso_lambda_max}
\title{Get the maximum lambda value for lasso}
\usage{
get_lasso_lambda_max(
  x,
  y,
  scalex = FALSE,
  nlambda = 100,
  lambda_min_ratio = 1e-04
)
}
\arguments{
\item{lambda_min_ratio}{}
}
\value{
lambda_max
}
\description{
Get the maximum lambda value for lasso
}
