% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_weight.R
\name{lasso_weight_single}
\alias{lasso_weight_single}
\title{Implement weighted Lasso estimation (second step of adaptive lasso) via coordinate descent
for the case in which only one predictor remains}
\usage{
lasso_weight_single(x, y, lambda, w = NULL, intercept = TRUE,
  scalex = FALSE)
}
\arguments{
\item{x}{Predictor matrix (n-by-p matrix)}

\item{y}{Response variable}

\item{lambda}{Shrinkage tuning parameter}

\item{w}{weights}

\item{intercept}{A boolean: include an intercept term or not}

\item{scalex}{A boolean: standardize the design matrix or not}
}
\value{
A list contains estimated intercept and slope
\item{ahat}{Estimated intercept}
\item{bhat}{Estimated slope}
}
\description{
Estimation function. Tuning parameter inputs needed.\cr
}
