% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrcov.R
\name{lrcov_est}
\alias{lrcov_est}
\title{Long-run covariance estimator}
\usage{
lrcov_est(x, y = NULL, type = 0)
}
\arguments{
\item{y}{= NULL if computing the long run variance of x only}

\item{type}{= 0 (default) for long-run covariance or 1 for one-sided long-run variance}
}
\description{
Use bartlett kernel to estimate long-run covariance where the bandwidth
determined Andrews (1991, Eq. (6.2) and (6.4)).
Part of the function adapted from the "bwAndrews" function in
"sandwich" package
}
