% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_coef}
\alias{plot_coef}
\title{Coefficient plot}
\usage{
plot_coef(coef_est, dates, pt_num = 4, col_vec = NULL,
  line_size = rep(0.75, length(coef_est)), alpha_size = rep(0.75,
  length(coef_est)), xlab = NULL, ylab = NULL, num_col = 1)
}
\arguments{
\item{coef_est}{estimated slope by m different methods, list of length m with each element n-by-p matrix\cr
names of list well defined}

\item{dates}{Date vector of class "Date". \cr}

\item{pt_num}{Number date points (pt_num + 1) wanted on the x-axis
Use \textit{lubridate} and \textit{zoo} to transfer original date vector to "Date" class}

\item{col_vec}{A vector indicates colors of different trends. \cr
e.g.:  c("#D8DBE2", "#F46B7B", "#518DE8", "#FFBC42" )\cr
If NULL, use ggplot default.}

\item{line_size}{line size for each trend}

\item{alpha_size}{degree of appearance}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
A ggplot output
}
\description{
Coefficient plot
}
\examples{
data("forecast_result")
data("raw_data_h1")
D <- raw_data_h1[!is.na(raw_data_h1$LongReturn), ]
dates <- zoo::as.Date.yearmon(D$yyyymm[-(1:180)])

coef_est <- list(lasso = forecast_result$Lasso$beta_hat[, 1:6],
                 alasso = forecast_result$ALasso$beta_hat[, 1:6])
plot_coef(coef_est, dates)

}
