% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_trend}
\alias{plot_trend}
\title{Trend plot}
\usage{
plot_trend(y_0, y_hat, dates, pt_num = 4, col_vec = NULL,
  line_size = rep(0.75, ncol(y_hat) + 1), alpha_size = rep(0.75,
  ncol(y_hat) + 1), xlab = NULL, ylab = NULL)
}
\arguments{
\item{y_0}{True predict target, length n vector}

\item{y_hat}{Predicted values by m different methods, n-by-m matrix, colnames well defined}

\item{dates}{Date vector of class "Date". \cr
Use \textit{lubridate} and \textit{zoo} to transfer original date vector to "Date" class}

\item{pt_num}{Number date points (pt_num + 1) wanted on the x-axis}

\item{col_vec}{A vector indicates colors of different trends. \cr
e.g.:  c("#D8DBE2", "#F46B7B", "#518DE8", "#FFBC42" )\cr
If NULL, use ggplot default.}

\item{line_size}{line size for each trend}

\item{alpha_size}{degree of appearance}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
A ggplot output
}
\description{
Trend plot
}
\examples{
data("forecast_result")
data("raw_data_h1")
D <- raw_data_h1[!is.na(raw_data_h1$LongReturn), ]
dates <- zoo::as.Date.yearmon(D$yyyymm[-(1:180)])
y0 <- forecast_result$y[-(1:180)]
y_hat <-cbind(forecast_result$Lasso$y_hat,
              forecast_result$Lasso_Std$y_hat,
              forecast_result$ALasso$y_hat)
colnames(y_hat) <- c("lasso", "lasso_std", "alasso")
plot_trend(y0, y_hat, dates)

}
