% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postlasso.R
\name{post_lasso}
\alias{post_lasso}
\title{post-selection estimation}
\usage{
post_lasso(x, y, coef.est, intercept = TRUE, scalex = FALSE)
}
\arguments{
\item{x}{Predictor matrix (n-by-p matrix)}

\item{y}{Response variable}

\item{coef.est}{shrinkage estimation results}

\item{intercept}{A boolean: include an intercept term or not}
}
\value{
A list contains estimated intercept and slope
\item{ahat}{Estimated intercept}
\item{bhat}{Estimated slope}
}
\description{
post-selection estimation
}
\examples{
post_lasso(x,y,coef.est)

}
