% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_predict.R
\name{roll_predict}
\alias{roll_predict}
\title{Rolling window forecast}
\usage{
roll_predict(x, y, roll_window, h = 1, methods_use = c("RW", "RWwD",
  "OLS", "Lasso", "Lasso_Std", "ALasso", "RepLasso", "post_Lasso",
  "post_Lasso_Std", "post_ALasso", "post_RepLasso"), train_method_las,
  verb = TRUE)
}
\arguments{
\item{x}{Full sample predictor}

\item{y}{Full sample forecast target}

\item{roll_window}{Length of the rolling window}

\item{h}{forecast horizon}

\item{train_method_las}{parameter tuning method for Lasso type methods. For comparison, all Lasso type methods shares the same train_method.}

\item{method_use}{method in use}
}
\description{
The function reads data and make forecasts based on linear predictive regression
with diverse methods. It incorporates both short-horizon and long-horizon forecasting.
}
