% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_talasso}
\alias{train_talasso}
\title{Do parameter tuning for replasso}
\usage{
train_talasso(
  x,
  y,
  b_first,
  gamma = 1,
  intercept = TRUE,
  scalex = FALSE,
  train_method = "timeslice",
  lambda_seq = NULL,
  nlambda = 100,
  lambda_min_ratio = 1e-04,
  k = 10,
  initial_window = ceiling(nrow(x) * 0.7),
  horizon = 1,
  fixed_window = TRUE
)
}
\arguments{
\item{x}{Predictor matrix (n-by-p matrix)}

\item{y}{Response variable}

\item{b_first}{estimated slope from first step alasso}

\item{gamma}{Parameter controlling the inverse of first step estimate. By default = 1.}

\item{intercept}{A boolean: include an intercept term or not}

\item{scalex}{A boolean: standardize the design matrix or not}

\item{train_method}{"timeslice", "cv",  "aic", "bic", "aicc", "hqc"}

\item{lambda_seq}{Candidate sequnece of parameters. If NULL, the function generates the sequnce.}

\item{nlambda}{# of lambdas}

\item{lambda_min_ratio}{# lambda_min_ratio * lambda_max = lambda_min}

\item{k}{k-fold cv if "cv" is chosen}

\item{initial_window}{control "timeslice"}

\item{horizon}{control "timeslice"}

\item{fixed_window}{control "timeslice"}
}
\value{
bestTune
}
\description{
If all variables are killed in the first step: return a random number\cr
If more than 1 variables are left: just repeat the training process for alasso\cr
If only 1 variable remained: use a brute-force process do the cross-validation.\cr
FOR FUTURE WORK: INCORPORATE REPLASSO INTO CARET FRAMEWORK.
}
\examples{
train_talasso(x,y)
}
